chore: 응집도를 높이기 위한 폴더 구조 변경
