+++ src/components/CardPreview/constants/index.ts
@@ -1,0 +1,10 @@
+export const CARD_TYPE_PATH = {
+  VISA: "./assets/Visa.png",
+  MASTERCARD: "./assets/Mastercard.png",
+  NONE: "",
+};
+
+export const CARD_TYPE_PREFIXES = {
+  VISA: ["4"],
+  MASTERCARD: ["51", "52", "53", "54", "55"],
+};
