+++ src/components/InputField/Expiration/constants/index.ts
@@ -1,0 +1,9 @@
+export const EXPIRATION_DATE = {
+  MONTH: {
+    MIN: 1,
+    MAX: 12,
+  },
+  YEAR: {
+    MAX_OFFSET: 10,
+  },
+};
