+++ src/utils/hasInvalidSegment.ts
